backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# WebSocket & Socket.IO
python-socketio==5.11.0
websockets==12.0

# Scientific computing (for object detection)
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# ML model support
torch>=2.0.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp>=3.8.0
